Refactor view structure and add new components
- Replaced `debug_view` module with a new `components` module.
- Created `debug_view`, `models_view`, `settings_view`, and `whisper_view` components.
- Implemented `ModelsView` with functionality to fetch and download models, including progress tracking.
- Updated Tailwind CSS configuration to include custom primary colors and Flowbite plugin.

+++ src-tauri/src/manager.rs
@@ -39,7 +39,7 @@ impl TranscriptionManager {
     pub async fn download_model(
         &self,
         model_name: &str,
-        progress_callback: impl Fn(f32) + Send + Sync + 'static,
+        progress_callback: impl Fn(i32) + Send + Sync + 'static,
     ) -> Result<()> {
         let models = self.config.get_available_models();
         let model = models
